% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareBars.R
\name{compareBars}
\alias{compareBars}
\title{Function for creating simple comparative bar charts.}
\usage{
compareBars(data, ordinalVar = NULL, compareVar1 = NULL,
  compareVar2 = NULL, width = NULL, height = NULL,
  orientation = "vertical", xLabel = "", yLabel = "", titleLabel = "",
  subtitleLabel = "", compareVarFill1 = "#0072B2",
  compareVarFill2 = "#E69F00", minFillColor = "#ddd",
  fontFamily = "sans-serif", axisFormat = ".0s")
}
\arguments{
\item{data}{a data frame object in wide form. Must have one string/factor column, and two numeric
columns to compare.}

\item{ordinalVar}{unquoted name of the ordinal variable}

\item{compareVar1}{unquoted name of a numeric value to compare}

\item{compareVar2}{unquoted name of another numeric value to compare}

\item{width}{width of the chart in pixels}

\item{height}{height of the chart in pixels}

\item{orientation}{if "vertical", the bars will render vertically along the x-axis. If 'horizontal',
the bars will render horizontally along the y-axis.}

\item{xLabel}{optional label for the x-axis}

\item{yLabel}{optional label for the y-axis}

\item{titleLabel}{optional label for the title}

\item{subtitleLabel}{optional label for the subtitle}

\item{compareVarFill1}{fill color for one of the levels}

\item{compareVarFill2}{fill color for the other level}

\item{minFillColor}{fill color for the smaller value}

\item{fontFamily}{font family for the labels}

\item{axisFormat}{axis formatting option, cf. d3-format on github}
}
\description{
Simplify comparative bar charts with d3.js: two values, one bar. For each level of an ordinal variable,
compareBars compares the two values, and returns the smaller as a uniform color, and the larger
remainder as a distinct color.
}
\examples{
\dontrun{
library(gapminder)
library(tidyverse)
gapminder \%>\%
 group_by(continent, year) \%>\%
 summarize(populations = sum(pop)) \%>\%
 spread(continent, populations) \%>\%
 mutate(year = factor(year)) \%>\%
 compareBars(year, Americas, Europe)
}
}
